Run Verus queue task in outer container
