+++ examples/data-structures/repositories/queue-rs/.vibesys/tasks/verus-mpmc-open/accuracy/Cargo.toml.in
@@ -1,0 +1,8 @@
+[package]
+name = "vibesys-verus-mpmc-accuracy"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+queue-verus-mpmc = { path = __CANDIDATE_PATH__ }
